Lab2 uploaded complete versions

+++ CS6012/Week1/Day2/TimingExperiments/src/lab02/TimingExperiment02.java
@@ -6,8 +6,8 @@ public class TimingExperiment02 {
     long secondsToRun = 10;
     long startTime = System.currentTimeMillis();
     long lastTime = startTime;
-    int advanceCount = 0;
-    int loopCount = 0;
+    long advanceCount = 0;
+    long loopCount = 0;
     while (lastTime - startTime < 1000 * secondsToRun) {
       loopCount++;
       long currentTime = System.currentTimeMillis();
